added new controller classes for the respective games
